artem_branch merged

+++ src/routes/Post.js
@@ -1,6 +1,7 @@
 const express = require("express");
 const router = express.Router();
 const { schemaValidate, auth } = require("../middlewares");
+const { postValidate } = require("../validationSchemas");
 const { Post } = require("../models");
 const { Tag } = require("../models");
 
@@ -47,46 +48,53 @@ router.get("/", async (req, res) => {
   }
 });
 
-router.post("/", auth, async (req, res) => {
-  try {
-    const tagsArr = req.body.tags ? req.body.tags.split(", ") : [];
+router.post(
+  "/",
+  auth,
+  schemaValidate(postValidate.create),
+  async (req, res) => {
+    try {
+      const tagsArr = req.body.tags ? req.body.tags.split(", ") : [];
 
-    const existingTags = await Tag.find({
-      name: {
-        $in: tagsArr,
-      },
-    });
+      const existingTags = await Tag.find({
+        name: {
+          $in: tagsArr,
+        },
+      });
 
-    const filtredTags = tagsArr.filter((tagName) => {
-      return !existingTags.find((tag) => tag.name === tagName);
-    });
-    const newTags = await Tag.insertMany(
-      filtredTags.map((tagName) => ({
-        name: tagName,
-      }))
-    );
+      const filtredTags = tagsArr.filter((tagName) => {
+        return !existingTags.find((tag) => tag.name === tagName);
+      });
+      const newTags = await Tag.insertMany(
+        filtredTags.map((tagName) => ({
+          name: tagName,
+        }))
+      );
 
-    const newPost = await Post.create({
-      title: req.body.title,
-      tags: tagsArr ? [...existingTags, ...newTags] : [],
-      // author: req.user._id,
-      thumbnailUrl: req.body.thumbnailUrl,
-      usersReading: req.body.usersReading,
-      usersLiked: req.body.usersLiked,
-      views: req.body.views,
-      body: req.body.body,
-    });
+      const newPost = await Post.create({
+        title: req.body.title,
+        tags: tagsArr ? [...existingTags, ...newTags] : [],
+        author: req.user._id,
+        thumbnailUrl: req.body.thumbnailUrl,
+        usersReading: req.body.usersReading,
+        usersLiked: req.body.usersLiked,
+        views: req.body.views,
+        body: req.body.body,
+      });
 
-    res.json({ newTags, newPost });
-  } catch (error) {
-    console.log(error);
-    res.status(500).send(error);
+      res.json({ newTags, newPost });
+    } catch (error) {
+      console.log(error);
+      res.status(500).send(error);
+    }
   }
-});
+);
 
 router.get("/:_id", async (req, res) => {
   try {
-    const searchedPost = await Post.findById(req.params._id);
+    const searchedPost = await Post.findById(req.params._id).populate(
+      "comments"
+    );
     res.json(searchedPost);
   } catch (error) {
     console.log(error);
@@ -94,20 +102,40 @@ router.get("/:_id", async (req, res) => {
   }
 });
 
-router.put("/:_id", async (req, res) => {
-  try {
-    const editedPost = await Post.findByIdAndUpdate(req.params._id, req.body, {
-      new: true,
-    });
-    res.json(editedPost);
-  } catch (error) {
-    console.log(error);
-    res.status(500).send(error);
-  }
-});
+router.put(
+  "/:_id",
+  auth,
+  schemaValidate(postValidate.update),
+  async (req, res) => {
+    try {
+      const post = await Post.findById(req.params._id);
 
-router.delete("/:_id", async (req, res) => {
+      if (post.author._id !== req.user._id) {
+        res.status(403).json({ message: "Error 403" });
+        return;
+      }
+      const editedPost = await Post.findByIdAndUpdate(
+        req.params._id,
+        req.body,
+        {
+          new: true,
+        }
+      );
+      res.json(editedPost);
+    } catch (error) {
+      console.log(error);
+      res.status(500).send(error);
+    }
+  }
+);
+
+router.delete("/:_id", auth, async (req, res) => {
   try {
+    const post = await Post.findById(req.params._id);
+    if (post.author._id !== req.user._id) {
+      res.status(403).json({ message: "Error 403" });
+      return;
+    }
     await Post.findByIdAndDelete(req.params._id);
     res.json({ message: "Done" });
   } catch (error) {
@@ -116,23 +144,45 @@ router.delete("/:_id", async (req, res) => {
   }
 });
 
-router.patch("/:_id/likes", async (req, res) => {
+router.patch("/:_id/likes", auth, async (req, res) => {
   try {
-    await Post.findOneAndUpdate(req.params._id.usersLiked, req.body, {
-      new: true,
-    });
-    res.json({ message: "Done" });
+    const post = await Post.findById(req.params._id);
+
+    if (req.user.likedPosts.includes(req.params._id)) {
+      post.usersLiked--;
+
+      req.user.likedPosts.pull(post._id);
+    } else {
+      post.usersLiked++;
+
+      req.user.likedPosts.push(post._id);
+    }
+    await post.save();
+    await req.user.save();
+
+    res.json(post);
   } catch (error) {
     console.log(error);
     res.status(500).send(error);
   }
 });
 
-router.patch("/:_id/save", async (req, res) => {
+router.patch("/:_id/save", auth, async (req, res) => {
   try {
-    await Post.findOneAndUpdate(req.params._id.usersReading, req.body, {
-      new: true,
-    });
+    const post = await Post.findById(req.params._id);
+
+    if (req.user.readingList.includes(req.params._id)) {
+      post.usersReading--;
+
+      req.user.readingList.pull(post._id);
+    } else {
+      post.usersReading++;
+
+      req.user.readingList.push(post._id);
+    }
+    await post.save();
+    await req.user.save();
+
     res.json({ message: "Done" });
   } catch (error) {
     console.log(error);
